Created a variable to manage currentQuestion
It creates a currentQuestion, then we run an if-statement to check if there is already a questionNumber on the session for us to then update from. If we don't have a questionNumber on the session then we set it as 1 since it would indicate that we are starting the survey.